requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
authlib
google-auth
google-auth-oauthlib
google-api-python-client
langchain
openai
python-dotenv
httpx
requests